Done task

+++ app/src/main/java/com/example/todo/data/network/entity/NewTask.kt
@@ -10,5 +10,5 @@ data class NewTask(
     @SerializedName("priority_id") val priority: Int,
     val done:Int = 0,
     val id: Int = 0,
-    val created: Int = 0
+    val created: Long = 0
 )
